requirements.txt
# Backend Framework
Django>=4.2,<6.1

# Deep Learning (CNN - Image Classification)

tensorflow>=2.12,<2.21
h5py>=3.8
Pillow>=9.5

# Scientific Computing

numpy>=1.23
scipy>=1.10
pandas>=1.5

# Machine Learning & NLP (Chatbot)

scikit-learn>=1.2
joblib>=1.3

# Visualization (Optional)

matplotlib>=3.7

#Commands to install the above libraries

#1.Create virtual Environment before running the project 
#python -m venv cattle-venv

#2.Activate the Virtual Environment
# venv/bin/activate

#3.Upgrade pip
#python -m pip install --upgrade pip

#4.Install all the libararies
#pip install django tensorflow numpy pandas scipy Pillow scikit-learn

#5.Install all the above things using this command
#pip install -r requirements.txt